add yaml files for remaining projects
